+++ Flipper/TradeData.cs
@@ -2,12 +2,12 @@
 {
     public class TradeData
     {
-        public int Amount { get; set; }
+        public decimal Amount { get; set; }
         public string Currency { get; set; }
         public string Name { get; set; }
         public string TypeLine { get; set; }
         public int Quantity { get; set; }
-        public int FullStackPrice { get; set; }
+        public decimal FullStackPrice { get; set; }
         public string Seller { get; set; }
     }
 }
